legacy agent + tables query refactor
